fix(plugins): remove blue ocean
Blue Ocean in the latest version of the Jenkins master image is failing with the following error when booting:

    java.lang.NoClassDefFoundError: io/jenkins/blueocean/commons/ServiceException$BadRequestExpception

Notice the typo in BadRequestExpception.

I made a really good attempt at bumping blue ocean to v1.1.6, but there appears to be a large amount of work required in wrangling the dependencies, so I did what any good developer would do and remove the plugin entirely. I like blue ocean, but considering that the CI infrastructure is going away within the next 6 months I would prefer stability over elegance.

That being said, if anyone is able to get Blue Ocean working again for CI, I'm more than happy to add it back! I just want to unblock CI at this point.